+++ frontend/component-lib/src/components/Container.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
     import {
+        ColourVars,
         getAlignmentCss,
         getBorderRadiusCss,
         getBorderStyleCss,
@@ -49,11 +50,16 @@
         colour?: string;
         mainAxisAlignment?: MainAxisAlignment;
         crossAxisAlignment?: CrossAxisAlignment;
+        mainAxisSelfAlignment?: MainAxisAlignment;
+        crossAxisSelfAlignment?: CrossAxisAlignment;
         minWidth?: Pixel;
         minHeight?: Pixel;
         shadow?: string;
         backgroundColour?: string;
         onClick?: () => void;
+        supplementalClass?: string;
+        allowOverflow?: boolean;
+        tag?: "div" | "main" | "section"; // this could be just about anything but let's try to limit it
     }
 
     let {
@@ -67,14 +73,19 @@
         height = { kind: "hug" },
         colour,
         borderStyle = "solid",
-        borderColour = "var(--background-2)",
+        borderColour = ColourVars.background2,
         mainAxisAlignment = "start",
         crossAxisAlignment = "start",
+        mainAxisSelfAlignment,
+        crossAxisSelfAlignment,
         minWidth = new Pixel(0),
         minHeight = new Pixel(0),
         shadow,
         backgroundColour = "unset",
         onClick,
+        supplementalClass,
+        allowOverflow = false,
+        tag = "div",
     }: Props = $props();
 
     // you might expect this to be done inside onMount but
@@ -90,7 +101,14 @@
     let widthCss = $derived(getFlexStyle("width", width, parentDirection));
     let heightCss = $derived(getFlexStyle("height", height, parentDirection));
     let colourCss = $derived(colour ? `background-color: ${colour}` : "");
-    let alignmentCss = $derived(getAlignmentCss(mainAxisAlignment, crossAxisAlignment));
+    let alignmentCss = $derived(
+        getAlignmentCss(
+            mainAxisAlignment,
+            crossAxisAlignment,
+            mainAxisSelfAlignment,
+            crossAxisSelfAlignment,
+        ),
+    );
     let style = $derived(
         `background-color: ${backgroundColour}; box-shadow: ${shadow}; min-width: ${minWidth}; min-height: ${minHeight}; ${alignmentCss}; ${colourCss}; ${heightCss}; ${widthCss}; ${borderStyleCss}; ${borderRadiusCss}; ${borderWidthCss}; ${paddingCss}; ${gapCss};`,
     );
@@ -100,13 +118,15 @@
 
 <!-- svelte-ignore a11y_click_events_have_key_events -->
 <!-- svelte-ignore a11y_no_static_element_interactions -->
-<div
+<svelte:element
+    this={tag}
     class:clickable={onClick !== undefined}
+    class:overflow={allowOverflow}
     onclick={onClick}
     {style}
-    class={`container ${direction}`}>
+    class={`container ${direction} ${supplementalClass ?? ""}`}>
     {@render children()}
-</div>
+</svelte:element>
 
 <style lang="scss">
     .container {
@@ -116,6 +136,10 @@
         transition:
             padding ease-in-out 200ms,
             gap ease-in-out 200ms;
+
+        &.overflow {
+            overflow: visible;
+        }
 
         &.horizontal {
             display: flex;