+++ frontend/component-lib/src/theme/avatars.ts
@@ -3,6 +3,7 @@ import { CssVariable } from "./variable";
 
 export class Avatars {
     constructor(
+        public xs: Pixel,
         public sm: Pixel,
         public md: Pixel,
         public lg: Pixel,
@@ -10,6 +11,7 @@ export class Avatars {
     ) {}
     cssVariables(): CssVariable[] {
         return [
+            this.xs.cssVariable("avatar", "xs"),
             this.sm.cssVariable("avatar", "sm"),
             this.md.cssVariable("avatar", "md"),
             this.lg.cssVariable("avatar", "lg"),