+++ frontend/component-test/src/App.svelte
@@ -4,7 +4,6 @@
     import MenuItem from "../../component-lib/src/components/menu/MenuItem.svelte";
     import Avatars from "./Avatars.svelte";
     import BigButtons from "./BigButtons.svelte";
-    import BottomBars from "./BottomBars.svelte";
     import Buttons from "./Buttons.svelte";
     import ChatSummaries from "./ChatSummaries.svelte";
     import Colours from "./Colours.svelte";
@@ -15,6 +14,7 @@
     import ListActions from "./ListActions.svelte";
     import Longpress from "./Longpress.svelte";
     import Menus from "./Menus.svelte";
+    import Messages from "./Messages.svelte";
     import SectionHeaders from "./SectionHeaders.svelte";
     import Typography from "./Typography.svelte";
     import Users from "./Users.svelte";
@@ -33,7 +33,7 @@
         | "longpress"
         | "controls"
         | "menus"
-        | "bottom_bar"
+        | "messages"
         | "avatars"
         | "section_header"
         | "typography"
@@ -44,7 +44,6 @@
     const labels: Record<Section, string> = {
         avatars: "Avatars",
         big_buttons: "Big buttons",
-        bottom_bar: "Bottom bar",
         buttons: "Buttons",
         chat_summary: "Chat summary",
         colours: "Colours",
@@ -55,6 +54,7 @@
         list_actions: "List Actions",
         longpress: "Longpress",
         menus: "Menus",
+        messages: "Messages",
         section_header: "Section headers",
         typography: "Typography",
         users: "Users",
@@ -100,8 +100,8 @@
             <ListActions></ListActions>
         {:else if selected === "menus"}
             <Menus></Menus>
-        {:else if selected === "bottom_bar"}
-            <BottomBars />
+        {:else if selected === "messages"}
+            <Messages></Messages>
         {:else if selected === "avatars"}
             <Avatars />
         {:else if selected === "forms"}