+++ backend/canisters/user_index/CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- sync streak data to local user indexes ([#8613](https://github.com/open-chat-labs/open-chat/pull/8613))
+
+## [[2.0.1903](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1903-user_index)] - 2025-09-12
+
+### Changed
+
+- Keep bot event subscriptions in sync ([#8600](https://github.com/open-chat-labs/open-chat/pull/8600))
+
 ## [[2.0.1895](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1895-user_index)] - 2025-09-11
 
 ### Changed